requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pillow==10.2.0
numpy<2.0
torch==2.2.0
torchvision==0.17.0
torchattacks==3.5.1
joblib==1.5.3
git+https://github.com/openai/CLIP.git
ultralytics==8.2.88
